requirements.txt
asgiref==3.4.1
astroid==2.9.0
autopep8==1.6.0
certifi==2021.10.8
charset-normalizer==2.0.9
coreapi==2.3.3
coreschema==0.0.4
Django==4.0
django-cors-headers==3.10.1
django-filter==21.1
djangorestframework==3.13.1
drf-yasg==1.20.0
idna==3.3
inflection==0.5.1
isort==5.10.1
itypes==1.2.0
Jinja2==3.0.3
lazy-object-proxy==1.7.1
MarkupSafe==2.0.1
mccabe==0.6.1
packaging==21.3
platformdirs==2.4.0
pycodestyle==2.8.0
pylint==2.12.2
pyparsing==3.0.6
pytz==2021.3
requests==2.26.0
ruamel.yaml==0.17.18
ruamel.yaml.clib==0.2.6
sqlparse==0.4.2
toml==0.10.2
typing-extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.7
wrapt==1.13.3
